requirements.txt
GitPython==2.1.3
PyYAML==3.12